+++ packages/iov-bcp/src/connection.ts
@@ -160,19 +160,19 @@ export function isFailedTransaction(
   return !isConfirmedTransaction(transaction);
 }
 
-export interface BcpQueryTag {
+export interface QueryTag {
   readonly key: string;
   readonly value: string;
 }
 
-export interface BcpTxQuery {
+export interface TransactionQuery {
   readonly id?: TransactionId;
   /** any send transaction to or from this address */
   readonly sentFromOrTo?: Address;
   /** any transaction signed by this address */
   readonly signedBy?: Address;
   /** chain-specific key value pairs that encode a query */
-  readonly tags?: ReadonlyArray<BcpQueryTag>;
+  readonly tags?: ReadonlyArray<QueryTag>;
   readonly height?: number;
   readonly minHeight?: number;
   readonly maxHeight?: number;
@@ -213,20 +213,22 @@ export interface BlockHeader {
   readonly transactionCount: number;
 }
 
-// BcpConnection is a high-level interface to a blockchain node,
-// abstracted over all blockchain types and communication channel.
-// A direct connection or a proxy server should implement this.
-// The implementation takes care to convert our internal types into
-// the proper format for the blockchain.
-//
-// BcpConnection is the minimal interface needed to be supported by any blockchain
-// that is compatible with the bcp spec and iov-core library. This supports
-// getting account balances, sending tokens, and observing the blockchain state.
-//
-// There are other optional interfaces that extend this functionality with
-// features like atomic swap, NFTs, etc which may be implemented by any connector
-// to enable enhanced features in the clients
-export interface BcpConnection {
+/**
+ * A high-level interface to a blockchain node,
+ * abstracted over all blockchain types and communication channel.
+ * A direct connection or a proxy server should implement this.
+ * The implementation takes care to convert our internal types into
+ * the proper format for the blockchain.
+ *
+ * BlockchainConnection is the minimal interface needed to be supported by any blockchain
+ * that is compatible with the BCP spec and IOV-Core library. This supports
+ * getting account balances, sending tokens, and observing the blockchain state.
+ *
+ * There are other optional interfaces that extend this functionality with
+ * features like atomic swap, NFTs, etc which may be implemented by any connector
+ * to enable enhanced features in the clients.
+ */
+export interface BlockchainConnection {
   // blockchain
   readonly disconnect: () => void;
   readonly chainId: () => ChainId;
@@ -262,16 +264,18 @@ export interface BcpConnection {
 
   // transactions
   readonly postTx: (tx: PostableBytes) => Promise<PostTxResponse>;
-  readonly searchTx: (query: BcpTxQuery) => Promise<ReadonlyArray<ConfirmedTransaction | FailedTransaction>>;
+  readonly searchTx: (
+    query: TransactionQuery,
+  ) => Promise<ReadonlyArray<ConfirmedTransaction | FailedTransaction>>;
   /**
    * Subscribes to all newly added transactions that match the query
    */
-  readonly listenTx: (query: BcpTxQuery) => Stream<ConfirmedTransaction | FailedTransaction>;
+  readonly listenTx: (query: TransactionQuery) => Stream<ConfirmedTransaction | FailedTransaction>;
   /**
    * Returns a stream for all historical transactions that match
    * the query, along with all new transactions arriving from listenTx
    */
-  readonly liveTx: (txQuery: BcpTxQuery) => Stream<ConfirmedTransaction | FailedTransaction>;
+  readonly liveTx: (query: TransactionQuery) => Stream<ConfirmedTransaction | FailedTransaction>;
   readonly getFeeQuote: (tx: UnsignedTransaction) => Promise<Fee>;
   // withDefaultFee will set the fee of the transaction to the result of getFeeQuote
   readonly withDefaultFee: <T extends UnsignedTransaction>(tx: T) => Promise<T>;