+++ packages/iov-bcp/types/chainconnector.d.ts
@@ -1,8 +1,8 @@
 import { TxCodec } from "./codec";
-import { BcpConnection } from "./connection";
+import { BlockchainConnection } from "./connection";
 import { ChainId } from "./transactions";
 export interface ChainConnector {
-    readonly client: () => Promise<BcpConnection>;
+    readonly client: () => Promise<BlockchainConnection>;
     readonly codec: TxCodec;
     readonly expectedChainId?: ChainId;
 }