+++ packages/iov-lisk/src/liskconnection.ts
@@ -8,9 +8,8 @@ import {
   AccountQuery,
   AddressQuery,
   Algorithm,
-  BcpConnection,
   BcpTicker,
-  BcpTxQuery,
+  BlockchainConnection,
   BlockHeader,
   BlockId,
   BlockInfo,
@@ -27,6 +26,7 @@ import {
   PublicKeyBytes,
   TokenTicker,
   TransactionId,
+  TransactionQuery,
   TransactionState,
   UnsignedTransaction,
 } from "@iov/bcp";
@@ -67,7 +67,7 @@ async function loadChainId(baseUrl: string): Promise<ChainId> {
   return responseBody.data.nethash;
 }
 
-export class LiskConnection implements BcpConnection {
+export class LiskConnection implements BlockchainConnection {
   public static async establish(baseUrl: string): Promise<LiskConnection> {
     const chainId = await loadChainId(baseUrl);
     return new LiskConnection(baseUrl, chainId);
@@ -325,7 +325,7 @@ export class LiskConnection implements BcpConnection {
     return Stream.create(producer);
   }
 
-  public async searchTx(query: BcpTxQuery): Promise<ReadonlyArray<ConfirmedTransaction>> {
+  public async searchTx(query: TransactionQuery): Promise<ReadonlyArray<ConfirmedTransaction>> {
     if (query.height || query.tags || query.signedBy) {
       throw new Error("Query by height, tags or signedBy not supported");
     }
@@ -343,11 +343,11 @@ export class LiskConnection implements BcpConnection {
     }
   }
 
-  public listenTx(_: BcpTxQuery): Stream<ConfirmedTransaction | FailedTransaction> {
+  public listenTx(_: TransactionQuery): Stream<ConfirmedTransaction | FailedTransaction> {
     throw new Error("Not implemented");
   }
 
-  public liveTx(query: BcpTxQuery): Stream<ConfirmedTransaction | FailedTransaction> {
+  public liveTx(query: TransactionQuery): Stream<ConfirmedTransaction | FailedTransaction> {
     if (query.height || query.tags || query.signedBy) {
       throw new Error("Query by height, tags or signedBy not supported");
     }